+++ Fabric/src/main/java/fuzs/puzzleslib/impl/PuzzlesLibFabric.java
@@ -1,12 +1,12 @@
 package fuzs.puzzleslib.impl;
 
-import fuzs.puzzleslib.core.CoreServices;
+import fuzs.puzzleslib.core.CommonFactories;
 import net.fabricmc.api.ModInitializer;
 
 public class PuzzlesLibFabric implements ModInitializer {
 
     @Override
     public void onInitialize() {
-        CoreServices.FACTORIES.modConstructor(PuzzlesLib.MOD_ID).accept(new PuzzlesLib());
+        CommonFactories.INSTANCE.modConstructor(PuzzlesLib.MOD_ID).accept(new PuzzlesLib());
     }
 }