+++ Fabric/src/main/java/fuzs/puzzleslib/mixin/client/accessor/AbstractContainerScreenAccessor.java
@@ -1,6 +1,7 @@
 package fuzs.puzzleslib.mixin.client.accessor;
 
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
+import net.minecraft.world.inventory.Slot;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
@@ -18,4 +19,7 @@ public interface AbstractContainerScreenAccessor {
 
     @Accessor("topPos")
     int getGuiTop();
+
+    @Accessor("hoveredSlot")
+    Slot getSlotUnderMouse();
 }