+++ Forge/src/main/java/fuzs/puzzleslib/client/gui/screens/ForgeScreens.java
@@ -6,6 +6,8 @@ import net.minecraft.client.gui.components.Widget;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.client.renderer.entity.ItemRenderer;
+import net.minecraft.world.inventory.Slot;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 import java.util.Objects;
@@ -40,6 +42,7 @@ public final class ForgeScreens implements CommonScreens {
 
     @Override
     public List<Widget> getRenderableButtons(Screen screen) {
+        Objects.requireNonNull(screen, "Screen cannot be null");
         return screen.renderables;
     }
 
@@ -65,5 +68,11 @@ public final class ForgeScreens implements CommonScreens {
     public int getTopPos(AbstractContainerScreen<?> screen) {
         Objects.requireNonNull(screen, "Screen cannot be null");
         return screen.getGuiTop();
+    }
+
+    @Override
+    public @Nullable Slot getHoveredSlot(AbstractContainerScreen<?> screen) {
+        Objects.requireNonNull(screen, "Screen cannot be null");
+        return screen.getSlotUnderMouse();
     }
 }