+++ Forge/src/main/java/fuzs/puzzleslib/impl/PuzzlesLibForge.java
@@ -1,6 +1,6 @@
 package fuzs.puzzleslib.impl;
 
-import fuzs.puzzleslib.core.CoreServices;
+import fuzs.puzzleslib.core.CommonFactories;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLConstructModEvent;
@@ -11,6 +11,6 @@ public class PuzzlesLibForge {
 
     @SubscribeEvent
     public static void onConstructMod(final FMLConstructModEvent evt) {
-        CoreServices.FACTORIES.modConstructor(PuzzlesLib.MOD_ID).accept(new PuzzlesLib());
+        CommonFactories.INSTANCE.modConstructor(PuzzlesLib.MOD_ID).accept(new PuzzlesLib());
     }
 }